refactor: move indexing_performance to graph_gateway
