+++ Cargo.toml
@@ -24,6 +24,7 @@ graphql = { git = "https://github.com/edgeandnode/toolshed", tag = "graphql-v0.3
 headers = "0.4.0"
 hex = "0.4"
 indexer-selection = { git = "https://github.com/edgeandnode/candidate-selection", rev = "1d6db83" }
+parking_lot = "0.12.3"
 primitive-types = "0.12.2"
 rand = { version = "0.8", features = ["small_rng"] }
 receipts = { git = "https://github.com/edgeandnode/receipts", rev = "e94e0f1" }
