+++ gateway-framework/src/lib.rs
@@ -5,6 +5,7 @@ pub mod chain;
 pub mod chains;
 pub mod config;
 pub mod errors;
+pub mod exchange_rate;
 pub mod graphql;
 pub mod http;
 pub mod indexing;
@@ -12,6 +13,5 @@ pub mod ip_blocker;
 pub mod json;
 pub mod logging;
 pub mod metrics;
-pub mod network;
 pub mod scalar;
 pub mod utils;
