--- gateway-framework/src/network/mod.rs
@@ -1,2 +1,0 @@
-pub mod exchange_rate;
-pub mod indexing_performance;
