+++ graph-gateway/Cargo.toml
@@ -24,6 +24,7 @@ ipnetwork = { version = "0.20.0", default-features = false }
 itertools = "0.13.0"
 num-traits = "0.2.18"
 ordered-float = "4.2.0"
+parking_lot.workspace = true
 prometheus = { version = "0.13", default-features = false }
 prost = "0.12.4"
 rand.workspace = true
