+++ graph-gateway/src/client_query/context.rs
@@ -1,13 +1,13 @@
 use alloy_sol_types::Eip712Domain;
-use gateway_framework::{
-    budgets::Budgeter, chains::Chains, network::indexing_performance::IndexingPerformance,
-    scalar::ReceiptSigner,
-};
+use gateway_framework::{budgets::Budgeter, chains::Chains, scalar::ReceiptSigner};
 use ordered_float::NotNan;
 use tokio::sync::{mpsc, watch};
 use url::Url;
 
-use crate::{indexer_client::IndexerClient, network::NetworkService, reports};
+use crate::{
+    indexer_client::IndexerClient, indexing_performance::IndexingPerformance,
+    network::NetworkService, reports,
+};
 
 #[derive(Clone)]
 pub struct Context {
