+++ graph-gateway/src/indexing_performance.rs
