+++ graph-gateway/src/lib.rs
@@ -2,6 +2,7 @@ pub mod block_constraints;
 pub mod client_query;
 pub mod indexer_client;
 pub mod indexers;
+pub mod indexing_performance;
 pub mod network;
 pub mod reports;
 pub mod subgraph_studio;
